fixed security issues

+++ application.py
@@ -76,62 +76,73 @@ def post_content(postid):
 
 @app.route('/post/list', methods=['GET'])
 def post_list():
-    return c_articles.objects().order_by('-created_at').all().to_json()
+    if 'admin' in session:
+        return c_articles.objects().order_by('-created_at').all().to_json()
+    return redirect(url_for('index'))
 
 
 @app.route('/post/save', methods=['POST'])
 def post_save():
-    rawcontent = request.data
-    content = json.loads(rawcontent)
-
-    c_articles(content['title'], content['category'], content['body'],
+    if 'admin' in session:
+        rawcontent = request.data
+        content = json.loads(rawcontent)
+        c_articles(content['title'], content['category'], content['body'],
                'green unlock').save()
-    return 'Success'
+        return 'success'
+    return redirect(url_for('index'))
 
 
 @app.route('/post/save/<postid>', methods=['POST'])
 def post_update(postid):
-    rawcontent = request.data
-    content = json.loads(rawcontent)
-    c_articles.objects(id=postid).update(set__title=content['title'],
+    if 'admin' in session:
+        rawcontent = request.data
+        content = json.loads(rawcontent)
+        c_articles.objects(id=postid).update(set__title=content['title'],
             set__category=content['category'], set__body=content['body'
             ])
-    return 'Post ' + postid + ' modified'
-
+        return 'Post ' + postid + ' modified'
+    return redirect(url_for('index'))
+    
 @app.route('/post/upload', methods=['GET', 'POST'])
 def post_upload():
-    if request.method == 'POST':
-        file = request.files['file']
-        if file and '.' in file.filename and file.filename.rsplit('.', 1)[1] in ALLOWED_EXTENSIONS:
-            filename = str(int(time.time()*100))+'.'+file.filename.rsplit('.', 1)[1]
-            print filename
-            file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
-            return app.config['UPLOAD_FOLDER']+'/'+filename
-    return '''
+    if 'admin' in session:
+        if request.method == 'POST':
+            file = request.files['file']
+            if file and '.' in file.filename and file.filename.rsplit('.', 1)[1] in ALLOWED_EXTENSIONS:
+                filename = str(int(time.time()*100))+'.'+file.filename.rsplit('.', 1)[1]
+                file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
+                return app.config['UPLOAD_FOLDER']+'/'+filename
+            return '''
     <form action="" method=post enctype=multipart/form-data>
       <p><input type=file name=file>
          <input type=submit value=Upload>
     </form>
     '''
+    return redirect(url_for('index'))
 
 @app.route('/post/delete/<postid>')
 def post_delete(postid):
-    c_articles.objects(id=postid).first().delete()
-    return 'Post ' + postid + ' deleted'
+    if 'admin' in session:
+        c_articles.objects(id=postid).first().delete()
+        return 'Post ' + postid + ' deleted'
+    return redirect(url_for('index'))
 
 
 @app.route('/post/lock/<postid>')
 def post_lock(postid):
-    c_articles.objects(id=postid).first().update(set__lock='black lock')
-    return 'success'
+    if 'admin' in session:
+        c_articles.objects(id=postid).first().update(set__lock='black lock')
+        return 'success'
+    return redirect(url_for('index'))
 
 
 @app.route('/post/unlock/<postid>')
 def post_unlock(postid):
-    c_articles.objects(id=postid).first().update(set__lock='green unlock'
+    if 'admin' in session:
+        c_articles.objects(id=postid).first().update(set__lock='green unlock'
             )
-    return 'success'
-
+        return 'success'
+    return redirect(url_for('index'))
 
 @app.route('/search/json', methods=['POST'])
 def search_json():
